move .spawn() into helium::node module
