+++ src/net/server.rs
@@ -49,8 +49,6 @@ use mio::net as mio_net;
 
 use util::get_epoch_time_secs;
 
-use std::{thread, thread::JoinHandle};
-
 pub struct HttpPeer {
     pub network_id: u32,
     pub chain_view: BurnchainView,
@@ -87,33 +85,6 @@ impl HttpPeer {
             burnchain: burnchain,
             connection_opts: conn_opts
         }
-    }
-
-    // This function connects to the burndb, peerdb, and stacks chainstate dbs
-    //   _it does not call connect()_
-    pub fn spawn(mut self, my_addr: &SocketAddr, mut peerdb: PeerDB,
-                 burn_db_path: String, stacks_chainstate_path: String, 
-                 is_mainnet: bool, stacks_chain_id: u32,
-                 poll_timeout: u64) -> Result<JoinHandle<()>, net_error> {
-        self.bind(my_addr, 500)?;
-        let http_thread = thread::spawn(move || {
-            loop {
-                test_debug!("http wakeup");
-                let mut burndb = BurnDB::open(&burn_db_path, true)
-                    .expect("Error while instantiating burnchain db");
-                let mut chainstate = StacksChainState::open(
-                    is_mainnet, stacks_chain_id, &stacks_chainstate_path)
-                    .expect("Error while instantiating chainstate db");
-
-                let view = {
-                    let mut tx = burndb.tx_begin().unwrap();
-                    BurnDB::get_burnchain_view(&mut tx, &self.burnchain).unwrap()
-                };
-                self.run(view, &mut burndb, &mut peerdb, &mut chainstate, poll_timeout)
-                    .unwrap();
-            }
-        });
-        Ok(http_thread)
     }
 
     /// start serving
